futher lint fixes; annotated getters and constructors where possible

+++ src/main/java/com/hikari/hellofx/Entities/Connectable/Basic/ShapePane.java
@@ -9,11 +9,11 @@ class ShapePane extends StackPane{
 	private final Rectangle field;
 	private final Text label;
 	
-	public ShapePane(int size, Color color) {
+	public ShapePane(double size, Color color) {
 		this(size, "off", color);
 	}
 	
-	public ShapePane(int size, String text, Color color) {
+	public ShapePane(double size, String text, Color color) {
 		field = new Rectangle(0,0,size,size);
 		label = new Text(text);
 		field.setFill(color);
